[finance_report_vision] Split governance_report.py (1145 lines) into a package
Behaviour-preserving split of the SSOT governance report + incremental-gate
tool; public surface (main) preserved.

- _base.py: constants, gate metadata, yaml guard (_require_yaml).
- _types.py: GovernanceEntry / ManifestSource / GateViolation.
- _util.py: path/ref/family/proof + gate-target helpers.
- _metrics.py: coverage ratios, quality snapshot, trend comparison.
- _manifest.py: manifest entry loading + orphan detection.
- _report.py: report building.  _gate.py: the incremental gate.  cli.py: render + main.
- __init__: re-exports the full module surface (the test treats the module as a
  flat namespace) + `import subprocess` so `governance_report.subprocess.run`
  monkeypatch resolves.

Test-only: the `yaml` monkeypatch is repointed to the `_base` submodule (where
_require_yaml looks it up). 14 governance tests pass; the live gate wrapper
(tools/report_ssot_governance.py) runs clean; full collection clean.
